+++ ios/Classes/IboxproFlutterPlugin.h
@@ -1,26 +1,4 @@
 #import <Flutter/Flutter.h>
-#import "Account.h"
-#import "Acquirer.h"
-#import "APIAuthenticationResult.h"
-#import "APIHistoryResult.h"
-#import "APIResult.h"
-#import "BTDevice.h"
-#import "Card.h"
-#import "DescriptionProduct.h"
-#import "DescriptionProductField.h"
-#import "ExternalPayment.h"
-#import "PaymentContext.h"
-#import "PaymentController.h"
-#import "PaymentOption.h"
-#import "Purchase.h"
-#import "QRData.h"
-#import "RecurrentPaymentContext.h"
-#import "ReversePaymentContext.h"
-#import "SignatureView.h"
-#import "StepItem.h"
-#import "TransactionData.h"
-#import "TransactionItem.h"
-#import "TransactionProduct.h"
 
 @interface IboxproFlutterPlugin : NSObject<FlutterPlugin>
 @end