+++ lib/iboxpro_flutter.dart
@@ -1,13 +1,4 @@
-import 'dart:async';
+library iboxpro_flutter;
 
-import 'package:flutter/services.dart';
-
-class IboxproFlutter {
-  static const MethodChannel _channel =
-      const MethodChannel('iboxpro_flutter');
-
-  static Future<String> get platformVersion async {
-    final String version = await _channel.invokeMethod('getPlatformVersion');
-    return version;
-  }
-}
+export 'src/payment.dart';
+export 'src/types.dart';